requirements.txt
numpy 
pandas
streamlit
openai
langchain
plotly.express